Secure Input 検出時のユーザー通知を追加（設定でオンオフ可、デフォルトオフ）
Secure Input wedge は akaza.log と入力メニューの警告だけでは気づきにくい
(2026-07-23 の再発時、警告は11回記録されていたが気づけず Mac 再起動に至った)。
検出時に通知センターへ保持アプリ名つきの警告を出す。

- SecureInputNotifier: activateServer / didWake の既存検出ポイントから呼ぶ。
  有効化エピソードごとに1回だけ通知(activateServer 連発でもスパムしない)。
  ロック画面 (loginwindow) の正当な保持は通知しない。
- 設定 notifyOnSecureInput (デフォルトオフ)。設定画面のチェックボックスを
  オンにした時点で通知許可をリクエストする。
- GeneralSettingsView が swiftlint の行数制限を超えたため、辞書セクションを
  GeneralSettingsView+Dict.swift に切り出し(コード移動のみ)。

EnableSecureEventInput を10秒間握るシミュレーションで通知表示を実機確認済み。
